Improve connection reading api

+++ URocket/Utils/ReadResult.cs
@@ -10,12 +10,12 @@ namespace URocket.Utils;
 
 public readonly struct ReadResult
 {
-    public readonly int TailSnapshot;   // drain boundary
+    public readonly long TailSnapshot;   // drain boundary
     public readonly bool IsClosed;      // socket closed / connection returned
     public readonly int Error;          // 0 or -errno (optional)
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public ReadResult(int tailSnapshot, bool isClosed, int error = 0)
+    public ReadResult(long tailSnapshot, bool isClosed, int error = 0)
     {
         TailSnapshot = tailSnapshot;
         IsClosed     = isClosed;
